requirements.txt
anyio==4.2.0
# beautifulsoup4==4.12.3
# bs4==0.0.2
certifi==2023.11.17
# charset-normalizer==3.3.2
# cryptography==42.0.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
# numpy==1.26.4
# pandas==2.2.0
pillow==10.2.0
# pycparser==2.21
# PyMySQL==1.1.0
# python-dateutil==2.8.2
python-telegram-bot==20.7
# pytz==2024.1
requests==2.31.0
# six==1.16.0
sniffio==1.3.0
# soupsieve==2.5
typing_extensions==4.9.0
# tzdata==2024.1
# urllib3==2.2.0